+++ lib/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @adeficior/data-modifier
 
+## 2.0.2
+
+### Patch Changes
+
+- 94bb070: don't load optional modules & more codegen options
+- Updated dependencies [94bb070]
+  - @adeficior/data-modifier-loader@2.1.0
+  - @adeficior/data-modifier-core@2.0.2
+  - @adeficior/data-modifier-ingredients@2.0.2
+  - @adeficior/data-modifier-loot@2.0.2
+  - @adeficior/data-modifier-recipes@2.0.2
+  - @adeficior/data-modifier-tags@2.0.2
+
 ## 2.0.1
 
 ### Patch Changes