+++ modules/create/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @adeficior/data-modifier-create
 
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies [94bb070]
+  - @adeficior/data-modifier-core@2.0.2
+  - @adeficior/data-modifier-ingredients@2.0.2
+  - @adeficior/data-modifier-recipes@2.0.2
+  - @adeficior/data-modifier-tags@2.0.2
+
 ## 2.0.1
 
 ### Patch Changes